requirements.txt
# ─────────────────────────────────────────────
# Living API Docs — requirements.txt
# Tested compatible set for Python 3.10 / 3.11 / 3.12
# Install: pip install -r requirements.txt
# ─────────────────────────────────────────────

# UI
streamlit==1.38.0

# LangChain ecosystem — these three MUST be version-aligned
langchain==0.3.7
langchain-community==0.3.7
langchain-groq==0.2.1

# Groq SDK — must match langchain-groq requirement
groq==0.11.0

# Embeddings
sentence-transformers==3.1.1
langchain-huggingface==0.1.0

# Vector DB
faiss-cpu==1.8.0

# GitHub integration
gitpython==3.1.43
PyGithub==2.5.0
requests==2.31.0

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.1
pygments==2.17.0
